Changed type of priority value from uint32_t to uint16_t.

+++ hipd/modularization.h
@@ -15,7 +15,7 @@ int hip_register_handle_function(const uint32_t packet_type,
                                  int (*handle_function)(const uint32_t packet_type,
                                                         const uint32_t ha_state,
                                                         struct hip_packet_context *ctx),
-                                 const uint32_t priority);
+                                 const uint16_t priority);
 
 int hip_unregister_handle_function(const uint32_t packet_type,
                                    const uint32_t ha_state,
@@ -28,7 +28,7 @@ int hip_run_handle_functions(const uint32_t packet_type,
 void hip_uninit_handle_functions(void);
 
 int hip_register_maint_function(int (*maint_function)(void),
-                                const uint32_t priority);
+                                const uint16_t priority);
 
 int hip_run_maint_functions(void);
